Use dispatcher-native scalar types for transducer ops
Register the transducer joint and loss checked entry points directly with the dispatcher instead of routing through _dispatch adapters that immediately narrowed int64_t and double arguments back to int and float.

Keep the dispatcher-facing C++ signatures aligned with schema binding semantics, where schema int arrives as int64_t and schema float arrives as double. The only remaining narrowing is at CUDA kernel call sites that still require float values.

+++ apex/contrib/csrc/transducer/transducer_loss_kernel.cu
@@ -455,8 +455,8 @@ __global__ void transducer_loss_fused_vec_backward(const scalar_t* x, const scal
 }
 
 std::vector<at::Tensor> transducer_loss_cuda_forward(at::Tensor x, at::Tensor label, at::Tensor audLen,
-                                                        at::Tensor txtLen, at::Tensor batchOffset, int maxFLen,
-                                                        int blankIdx, int opt, bool packedInput) {
+                                                        at::Tensor txtLen, at::Tensor batchOffset, int64_t maxFLen,
+                                                        int64_t blankIdx, int64_t opt, bool packedInput) {
   auto scalarType = x.scalar_type();
   auto tensorOpt = x.options();
   const int batchSize = label.size(0);
@@ -515,8 +515,8 @@ std::vector<at::Tensor> transducer_loss_cuda_forward(at::Tensor x, at::Tensor la
 
 at::Tensor transducer_loss_cuda_backward(at::Tensor x, at::Tensor lossGrad, at::Tensor alpha,
                                             at::Tensor beta, at::Tensor audLen, at::Tensor txtLen,
-                                            at::Tensor label, at::Tensor batchOffset, int maxFLen, int blankIdx,
-                                            int opt, bool fuseSoftmaxBackward, bool packedInput) {
+                                            at::Tensor label, at::Tensor batchOffset, int64_t maxFLen,
+                                            int64_t blankIdx, int64_t opt, bool fuseSoftmaxBackward, bool packedInput) {
   auto dtype = x.scalar_type();
   at::Tensor xGrad;
   const int batchSize = label.size(0);
